Fix scripts to not hardcode istio/installer (#691)
We cannot be cloning from github blindly, as this will break our private
security builds. This PR adds an option to pass an INSTALLER_DIR option
to the scripts and skip the cloning. Additionally, do not run vfsgen
when you do docker build. This should be an explicit step, otherwise the
code magically changes during a build and you spend all day debugging
it.